components/soil/PCA_soil/requirements-dev.txt
pytest>=7.4.1
panda~=2.1.1